+++ CHANGELOG.md
@@ -10,6 +10,7 @@ API changes:
 * New `P.empty` predicate for checking empty lists/texts/etc.
 * Add `HasSubsequences` instance for lazy `Text`
 * Remove field prefixes from more constructors (`FlagSpec`, `TestResult`, `SpecInfo`)
+* Change `testResultSuccess` from `Bool` to `TestResultStatus`
 
 Runtime changes:
 * Sanitize a literal "\`\`\`" line in snapshots ([#27](https://github.com/brandonchinn178/skeletest/issues/27))